Move monitoring of contacts to a dedicated service
