+++ android_mobile/src/main/java/com/alexstyl/specialdates/AppComponent.java
@@ -5,6 +5,7 @@ import com.alexstyl.specialdates.addevent.AddEventActivity;
 import com.alexstyl.specialdates.addevent.EventDatePickerDialogFragment;
 import com.alexstyl.specialdates.addevent.ui.ContactSuggestionView;
 import com.alexstyl.specialdates.analytics.AnalyticsModule;
+import com.alexstyl.specialdates.contact.ContactMonitorService;
 import com.alexstyl.specialdates.contact.ContactsModule;
 import com.alexstyl.specialdates.dailyreminder.DailyReminderIntentService;
 import com.alexstyl.specialdates.dailyreminder.DailyReminderModule;
@@ -118,4 +119,6 @@ public interface AppComponent {
     void inject(ContactPermissionActivity activity);
 
     void inject(BootCompleteReceiver receiver);
+
+    void inject(ContactMonitorService contactMonitorService);
 }
